RF: Move icon test to test_components
