+++ psychopy/tests/test_app/test_builder/test_components.py
@@ -1,4 +1,7 @@
 from __future__ import print_function
+
+from pathlib import Path
+
 from past.builtins import unicode
 
 from builtins import object
@@ -164,6 +167,23 @@ class TestComponents(object):
 
         for mismatch in mismatched:
             warnings.warn("Non-identical Builder Param: {}".format(mismatch))
+
+    def test_icons(self):
+        """Check that all components have icons for each app theme"""
+        # Iterate through component classes
+        for comp in self.allComp.values():
+            if hasattr(comp, "icon"):
+                # If icon path specified, pathify it
+                icon = Path(comp.icon)
+                # Get paths for each theme
+                files = [
+                    icon.parent.parent / "light" / icon.name,
+                    icon.parent.parent / "dark" / icon.name,
+                    icon.parent.parent / "classic" / icon.name,
+                ]
+                # Check that each path is a file
+                for file in files:
+                    assert file.is_file()
 
 
 @pytest.mark.components
