+++ psychopy/tests/test_scripts/test_component_compile_python.py
@@ -77,20 +77,3 @@ class TestComponentCompilerPython(object):
                 # Check component in exp
                 component = compName.split('Component')[0]
                 assert self.exp.getComponentFromType(component)
-
-    def test_icons(self):
-        """Check that all components have icons for each app theme"""
-        # Iterate through component classes
-        for comp in self.allComp.values():
-            if hasattr(comp, "icon"):
-                # If icon path specified, pathify it
-                icon = Path(comp.icon)
-                # Get paths for each theme
-                files = [
-                    icon.parent.parent / "light" / icon.name,
-                    icon.parent.parent / "dark" / icon.name,
-                    icon.parent.parent / "classic" / icon.name,
-                ]
-                # Check that each path is a file
-                for file in files:
-                    assert file.is_file()
